+++ include/common/standard.h
@@ -61,7 +61,7 @@
  * only work with ints and booleans though.
  */
 #define likely(x) (x)
-#define unlikely(x) (__builtin_expect((x), 0))
+#define unlikely(x) (__builtin_expect((unsigned long)(x), 0))
 #endif
 #endif
